chore(native): rename transition manager and support more props
+ props documentation
