+++ packages/native/src/components/transitions/index.ts
@@ -1,4 +1,4 @@
 export * from "./Fade";
 export * from "./Slide";
 export * from "./types";
-export * from "./Manager";
+export * from "./Transition";
